+++ C++/graph/mst/kruskal.hpp
@@ -8,7 +8,7 @@ inline MST kruskal(std::vector<edge> edges, const int n) noexcept {
     std::ranges::sort(edges, [&](const edge &e, const edge &f) -> bool { return e.cost < f.cost; });
     UnionFind uf(n);
     std::vector<edge> e;
-    int64_t ret = 0;
+    long long ret = 0;
     for(const auto &ed: edges) {
         if(uf.unite(ed.src, ed)) {
             e.emplace_back(ed);
