+++ C++/math/ModPrime.hpp
@@ -14,11 +14,11 @@ template <class T> constexpr inline T mod(T x, const T m) noexcept {
 namespace man {
 template <int lim> struct ModPrime {
 private:
-    const int64_t m;
-	std::array<int64_t, lim> f{}, rf{};
-	const int len = std::min(m, (int64_t) lim);
-    constexpr inline int64_t inv(int64_t x) noexcept {
-        int64_t ret = 1, k = m - 2;
+    const long long m;
+	std::array<long long, lim> f{}, rf{};
+	const int len = std::min(m, (long long) lim);
+    constexpr inline long long inv(long long x) noexcept {
+        long long ret = 1, k = m - 2;
 		while(k > 0) {
 			if(k & 1) {
 				ret = mod(ret * x, m);
@@ -29,7 +29,7 @@ private:
 		return ret;
     }
 public:
-    ModPrime(const int64_t mod_): m(mod_) {
+    ModPrime(const long long mod_): m(mod_) {
 		f[0] = 1;
 		for(const auto i: std::views::iota(1, len)) {
 			f[i] = mod(f[i - 1] * i, m);
@@ -39,21 +39,21 @@ public:
 			rf[i - 1] = mod(rf[i] * i, m);
 		}
     }
-    constexpr inline int64_t C(const int n, const int k) const noexcept {
+    constexpr inline long long C(const int n, const int k) const noexcept {
 		if(k < 0 || n < k) {
 			return 0;
 		}
-		const int64_t a = f[n], b = rf[n - k], c = rf[k], bc = mod(b * c, m);
+		const long long a = f[n], b = rf[n - k], c = rf[k], bc = mod(b * c, m);
 		return mod(a * bc, m);
 	}
-	constexpr inline int64_t P(const int n, const int k) const noexcept {
+	constexpr inline long long P(const int n, const int k) const noexcept {
 		if(k < 0 || n < k) {
 			return 0;
 		}
-		const int64_t a = f[n], b = rf[n - k];
+		const long long a = f[n], b = rf[n - k];
 		return mod(a * b, m);
 	}
-	constexpr inline int64_t H(const int n, const int k) const noexcept {
+	constexpr inline long long H(const int n, const int k) const noexcept {
 		if(n == 0 && k == 0) {
 			return 1;
 		}
