+++ C++/math/inverse_num.hpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 #include "C++/ds/fwtree/FenwickTree.hpp"
 namespace man {
-template <std::integral T> constexpr inline int64_t inv_num(const std::vector<T> &a) noexcept {
+template <std::integral T> constexpr inline long long inv_num(const std::vector<T> &a) noexcept {
     const int n = std::ssize(a);
     std::vector<std::pair<T, int>> p(n);
     for(const auto i: std::views::iota(0, n)) {
@@ -15,7 +15,7 @@ template <std::integral T> constexpr inline int64_t inv_num(const std::vector<T>
         id[p[i].second] = i;
     }
     FenwickTree<T> bit(n);
-    int64_t ret = 0;
+    long long ret = 0;
     for(const auto i: std::views::iota(0, n)) {
         ret += i - bit.sum(id[i]);
         bit.add(id[i], 1);
