+++ C++/math/is_prime.hpp
@@ -1,8 +1,7 @@
 #pragma once
-#include <cstdint>
 
 namespace man {
-constexpr inline bool is_prime(const uint64_t n) noexcept {
+constexpr inline bool is_prime(const unsigned long long n) noexcept {
 	if(n <= 1) {
 		return false;
 	}
@@ -12,7 +11,7 @@ constexpr inline bool is_prime(const uint64_t n) noexcept {
 	if(n % 2 ==0 || n % 3 == 0) {
 		return false;
 	}
-	for(int64_t i = 5; i * i <= n; i += 6) {
+	for(long long i = 5; i * i <= n; i += 6) {
 		if(n % i == 0 || n % (i + 2) == 0) {
 			return false;
 		}
