+++ C++/other/doubling.hpp
@@ -4,7 +4,7 @@
 #include <cmath>
 #include <ranges>
 namespace man {
-template <class T> requires std::ranges::random_access_range<T> constexpr inline std::vector<std::vector<int>> doubling(const T &a, const int64_t k) noexcept {
+template <class T> requires std::ranges::random_access_range<T> constexpr inline std::vector<std::vector<int>> doubling(const T &a, const long long k) noexcept {
     const int z = std::ceil(std::log2(k)), n = std::ssize(a);
     std::vector dbl(z, std::vector(n, 0));
     for(const auto &i: std::views::iota(0, n)) {
